backend/requirements.txt
alembic==1.8.1
aniso8601==9.0.1
cffi==1.15.1
click==8.1.3
colorama==0.4.5
config==0.5.1
cryptography==38.0.1
Flask==2.2.2
Flask-JSON==0.3.4
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
passlib==1.7.4
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.5
six==1.16.0
SQLAlchemy==1.4.42
Werkzeug==2.2.2
zipp==3.9.0
